Add upstream's upcoming url tag
Django 1.2's url tag takes a constant as the view name that you're trying to
reverse. This is unfortunate for code that's trying to build a generic
"select a program" page that'll feed into various views. Django 1.3 fixes this
with a new version of the url tag in the "future" templatetag package (and
Django 1.5 will use a backwards-incompatible url tag). Thus, we're importing
Django 1.3's tag.